af: extract shared session helpers to _astro_session
discovery/astro_cli.py was importing _config_path and _read_yaml from
astro_pat. that couples discovery to auth-module internals: if astro_pat
later moves yaml parsing or grows a per-instance config root, discovery
breaks silently.

now: new module _astro_session holds the primitives both sides need
(_astro_home, _config_path, _read_yaml, _find_context, _context_key,
_parse_expiry, _bearer_from_ctx, _auth_config_url) and the shared base
exception AstroPATError + AstroNotLoggedInError. astro_pat re-exports
the names so existing test imports keep working. discovery now imports
from the stable shared location.

leading underscore on the module signals "internal to
astro-airflow-mcp" without exposing it as a public submodule.
